+++ monopyly/__init__.py
@@ -5,6 +5,7 @@ Run a development server for the Monopyly app.
 from flask import Flask
 
 from monopyly.config import DevelopmentConfig, ProductionConfig
+from monopyly.core.errors import render_error_template
 from monopyly.database import SQLAlchemy, register_db_cli_commands
 
 
@@ -32,6 +33,7 @@ def create_app(test_config=None):
 def init_app(app):
     """Initialize the app."""
     register_blueprints(app)
+    register_errorhandlers(app)
     register_db_cli_commands(app)
 
 
@@ -57,3 +59,20 @@ def register_blueprints(app):
     app.register_blueprint(auth_bp)
     app.register_blueprint(banking_bp)
     app.register_blueprint(credit_bp)
+
+
+def register_errorhandlers(app):
+    """Register error handlers with the app."""
+    handled_error_codes = [
+        400,
+        401,
+        403,
+        404,
+        405,
+        408,
+        418,
+        # 425 -- not yet supported
+        500,
+    ]
+    for code in handled_error_codes:
+        app.register_error_handler(code, render_error_template)