+++ CHANGELOG.md
@@ -175,5 +175,6 @@
 
 ### 1.4.7 (in progress)
 
+- Create custom error pages for error responses (400, 404, 500, etc.)
 - Refresh the table of transactions after making a payment on a credit card statement
 - Use SVG to handle long values in account/statement summary boxes; fixes bugs in page rendering (long value overflow) and hover actions not happening because of conflicting overlap with the sidebar